feat(core): emit OAuth login URL as a single OSC 8 hyperlink
The Qwen OAuth device-flow fallback message hard-wrapped the auth URL
across a fixed-width ASCII box, splitting it over multiple lines so it
couldn't be copied/clicked as one unit (notably over SSH). showFallbackMessage
now emits the URL as a single OSC 8 hyperlink when the terminal supports it
(gated by the existing supportsHyperlinks detector) and keeps the ASCII box
otherwise.

showFallbackMessage lives in core but the OSC 8 helpers lived in cli; extract
the shared primitives (sanitizeForOsc, osc8Hyperlink, supportsHyperlinks,
wrapForMultiplexer, HYPERLINK_ENV_KEYS) to packages/core/src/utils/osc8.ts and
re-export them from the existing cli modules -- one source of truth, no
duplication, all existing consumers unchanged. showFallbackMessage is exported
and takes an optional output stream (default process.stderr) for testability.

Fixes #6428.
